+++ common/design-compose/src/main/kotlin/app/ss/design/compose/theme/Type.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 import app.ss.design.compose.R
 
-private val LatoFontFamily = FontFamily(
+val LatoFontFamily = FontFamily(
     Font(R.font.lato_regular, FontWeight.Normal),
     Font(R.font.lato_medium, FontWeight.Medium),
     Font(R.font.lato_bold, FontWeight.Bold),
