--- common/design/src/main/kotlin/com/cryart/design/widgets/DragHandle.kt
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2021. Adventech <info@adventech.io>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.cryart.design.widgets
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-import com.cryart.design.theme.BaseGrey1
-import com.cryart.design.theme.BaseGrey2
-
-private object DragHandleDefaults {
-    val width = 48.dp
-    val height = 4.dp
-}
-
-@Composable
-fun DragHandle(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier
-            .size(
-                width = DragHandleDefaults.width,
-                height = DragHandleDefaults.height
-            )
-            .background(backgroundColor(), CircleShape)
-    )
-}
-
-@Composable
-private fun backgroundColor(): Color =
-    if (isSystemInDarkTheme()) {
-        BaseGrey2
-    } else {
-        BaseGrey1
-    }
