+++ features/lessons/src/main/java/com/cryart/sabbathschool/lessons/ui/quarterlies/model/QuarterlySpec.kt
@@ -26,8 +26,8 @@ import androidx.compose.runtime.Immutable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import app.ss.design.compose.theme.parse
 import app.ss.models.SSQuarterly
-import com.cryart.design.theme.parse
 
 @Immutable
 internal data class QuarterlySpec(
